add command to perform login and get user agent token
